Use shared pointers for managing Tileset instance lifetime
This was a rather tricky change to do, and I tried to do it both with a
QSharedPointer as well as QExplicitlySharedDataPointer like I used for
the TileStamp class. I even tried to use a std::shared_ptr for some
time, because I thought it could avoid depending on
QEnableSharedFromThis from Qt 5.4. I hope the current approach is the
right one in the end.

The idea is to simplify tileset lifetime management. Previously this was
done through the TilesetManager and MapDocument classes, but that made
working with plain Tiled::Map instances (like the tile stamp variations)
cumbersome. When you deleted a Tiled::Map, you could never really be
sure wether its tilesets should be deleted or not.

The TilesetManager is still around, since it also implements watching
the tileset images for changes.

Now that tilesets can be properly shared between maps, this should
eventually simplify the work required to add missing tilesets to the map
when you paint on it with some tiles from elsewhere.

The Python plugin is unfortunately disabled until it has been somehow
adapted to use these shared pointers.